ai_service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.12
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
idna==3.10
joblib==1.5.1
lightgbm==4.6.0
nltk==3.9.1
numpy==2.2.6
pandas==2.2.3
psycopg2==2.9.10
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
regex==2024.11.6
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
Unidecode==1.4.0
uvicorn==0.34.3
